refactor: move default objects to own namespace
- cluster common tab-data types
