+++ packages/editor/src/app-state.ts
@@ -1,6 +1,6 @@
-import { InscriptionDataBeta } from '@axonivy/inscription-protocol';
+import { InscriptionData } from '@axonivy/inscription-protocol';
 
-export type AppState = { state: 'waiting' } | { state: 'error'; error: string } | { state: 'success'; initialData: InscriptionDataBeta };
+export type AppState = { state: 'waiting' } | { state: 'error'; error: string } | { state: 'success'; initialData: InscriptionData };
 
 export function waitingState(): AppState {
   return { state: 'waiting' };
@@ -10,6 +10,6 @@ export function errorState(error: string): AppState {
   return { state: 'error', error };
 }
 
-export function successState(initialData: InscriptionDataBeta): AppState {
+export function successState(initialData: InscriptionData): AppState {
   return { state: 'success', initialData };
 }
