--- packages/protocol/src/data/end-page-data.ts
@@ -1,7 +1,0 @@
-export interface EndPageData {
-  page: string;
-}
-
-export const DEFAULT_END_PAGE_DATA: EndPageData = {
-  page: ''
-};
