+++ packages/protocol/src/data/index.ts
@@ -1,12 +1,7 @@
 export * from './inscription-data';
 export * from './inscription-type';
 
-export * from './name-data';
-export * from './call-data';
-export * from './output-data';
-export * from './task-data';
-export * from './case-data';
-export * from './end-page-data';
+export * from './tab-data';
+export * from './workflow-data';
 
-export * from './common';
 export * from './inscription';
