+++ packages/protocol/src/data/inscription-data.ts
@@ -1,17 +1,13 @@
-import { CallData, DEFAULT_CALL_DATA, DialogCallData, ProcessCallData } from './call-data';
-import { CaseData, DEFAULT_CASE_DATA } from './case-data';
-import { DEFAULT_END_PAGE_DATA, EndPageData } from './end-page-data';
-import { NameData, DEFAULT_NAME_DATA } from './name-data';
-import { DEFAULT_OUTPUT_DATA } from './output-data';
+import { CallData, DialogCallData, EndPageData, NameData, ProcessCallData } from './tab-data';
 import { ElementScript, InscriptionType } from './inscription';
-import { DEFAULT_TASK_DATA, TaskData } from './task-data';
 import { ElementType } from './inscription-type';
+import { CaseData, TaskData } from './workflow-data';
 
 export type ConfigData = CallData & DialogCallData & ProcessCallData & ElementScript & TaskData & CaseData & EndPageData;
 
 export type DataBeta = NameData & { config: ConfigData };
 
-export interface InscriptionDataBeta {
+export interface InscriptionData {
   pid: string;
   type: InscriptionType;
   readonly: boolean;
@@ -24,14 +20,3 @@ export interface InscriptionSaveData {
   type: ElementType;
   data: DataBeta;
 }
-
-export const DEFAULT_DATA: DataBeta = {
-  ...DEFAULT_NAME_DATA,
-  config: {
-    ...DEFAULT_CALL_DATA,
-    ...DEFAULT_OUTPUT_DATA,
-    ...DEFAULT_TASK_DATA,
-    ...DEFAULT_CASE_DATA,
-    ...DEFAULT_END_PAGE_DATA
-  }
-} as const;
