--- packages/protocol/src/data/output-data.ts
@@ -1,10 +1,0 @@
-import { ElementScript } from './inscription';
-
-
-export const DEFAULT_OUTPUT_DATA: ElementScript = {
-  output: {
-    map: {},
-    code: ''
-  },
-  sudo: false
-} as const;
