+++ packages/protocol/src/data/workflow-data.ts
@@ -1,0 +1,33 @@
+import { WfCase, WfTask } from './inscription';
+
+export const RESPONSIBLE_TYPE = {
+  ROLE: 'Role',
+  ROLE_FROM_ATTRIBUTE: 'Role from Attr.',
+  USER_FROM_ATTRIBUTE: 'User from Attr.',
+  DELETE_TASK: 'Nobody & delete'
+} as const;
+
+export const PRIORITY_LEVEL = {
+  LOW: 'Low',
+  NORMAL: 'Normal',
+  HIGH: 'High',
+  EXCEPTION: 'Exception',
+  SCRIPT: 'Script'
+} as const;
+
+export const CUSTOM_FIELD_TYPE = {
+  STRING: 'String',
+  TEXT: 'Text',
+  NUMBER: 'Number',
+  TIMESTAMP: 'Timestamp'
+} as const;
+
+export interface TaskData {
+  task: WfTask;
+  tasks: WfTask[];
+  persist: boolean;
+}
+
+export interface CaseData {
+  case: WfCase;
+}
