+++ packages/protocol/src/defaults/common-defaults.ts
@@ -1,0 +1,41 @@
+import { CallData, DataBeta, DialogCallData, Document, ElementScript, EndPageData, NameData, ProcessCallData } from '../data';
+import { DEFAULT_TASK_DATA, DEFAULT_CASE_DATA } from './workflow-defaults';
+
+export const DEFAULT_NAME_DATA: NameData = {
+  name: '',
+  description: '',
+  docs: [] as Document[],
+  tags: [] as string[]
+};
+
+export const DEFAULT_OUTPUT_DATA: ElementScript = {
+  output: {
+    map: {},
+    code: ''
+  },
+  sudo: false
+} as const;
+
+export const DEFAULT_CALL_DATA: CallData & DialogCallData & ProcessCallData = {
+  dialog: '',
+  processCall: '',
+  call: {
+    map: {},
+    code: ''
+  }
+} as const;
+
+export const DEFAULT_END_PAGE_DATA: EndPageData = {
+  page: ''
+};
+
+export const DEFAULT_DATA: DataBeta = {
+  ...DEFAULT_NAME_DATA,
+  config: {
+    ...DEFAULT_CALL_DATA,
+    ...DEFAULT_OUTPUT_DATA,
+    ...DEFAULT_TASK_DATA,
+    ...DEFAULT_CASE_DATA,
+    ...DEFAULT_END_PAGE_DATA
+  }
+} as const;
