+++ packages/protocol/src/index.ts
@@ -1,5 +1,6 @@
 export * from './inscription-client';
 export * from './inscription-protocol';
 export * from './data';
+export * from './defaults';
 export * from './meta';
 export * from './validation';
